+++ bank-feeds/src/codatbankfeeds/models/shared/__init__.py
@@ -19,6 +19,7 @@ from .dataconnectionerror import *
 from .dataconnectionstatus import *
 from .datatype import *
 from .halref import *
+from .items import *
 from .links import *
 from .pushchangetype import *
 from .pushoperation import *
@@ -33,4 +34,4 @@ from .validation import *
 from .validationitem import *
 from .zero import *
 
-__all__ = ["BankAccountCredentials","BankFeedAccountMappingResponse","BankFeedMapping","BankTransactionType","BankTransactions","ClientRateLimitReachedWebhook","ClientRateLimitReachedWebhookData","ClientRateLimitResetWebhook","ClientRateLimitResetWebhookData","Companies","Company","CompanyRequestBody","Connection","Connections","CreateBankTransactions","CreateBankTransactionsResponse","DataConnectionError","DataConnectionStatus","DataType","HalRef","Links","PushChangeType","PushOperation","PushOperationChange","PushOperationRef","PushOperationStatus","PushOperations","Security","SourceAccount","SourceType","TargetAccountOption","Validation","ValidationItem","Zero"]
+__all__ = ["BankAccountCredentials","BankFeedAccountMappingResponse","BankFeedMapping","BankTransactionType","BankTransactions","ClientRateLimitReachedWebhook","ClientRateLimitReachedWebhookData","ClientRateLimitResetWebhook","ClientRateLimitResetWebhookData","Companies","Company","CompanyRequestBody","Connection","Connections","CreateBankTransactions","CreateBankTransactionsResponse","DataConnectionError","DataConnectionStatus","DataType","GroupReference","HalRef","Items","Links","PushChangeType","PushOperation","PushOperationChange","PushOperationRef","PushOperationStatus","PushOperations","Security","SourceAccount","SourceType","TargetAccountOption","Validation","ValidationItem","Zero"]