+++ bank-feeds/src/codatbankfeeds/source_accounts.py
@@ -3,7 +3,7 @@
 from .sdkconfiguration import SDKConfiguration
 from codatbankfeeds import utils
 from codatbankfeeds.models import errors, operations, shared
-from typing import Optional
+from typing import List, Optional
 
 class SourceAccounts:
     r"""Source accounts act as a bridge to bank accounts in accounting software."""
@@ -305,8 +305,8 @@ class SourceAccounts:
         
         if http_res.status_code == 200:
             if utils.match_content_type(content_type, 'application/json'):
-                out = utils.unmarshal_json(http_res.text, Optional[shared.SourceAccount])
-                res.source_account = out
+                out = utils.unmarshal_json(http_res.text, Optional[List[shared.SourceAccount]])
+                res.source_accounts = out
             else:
                 raise errors.SDKError(f'unknown content-type received: {content_type}', http_res.status_code, http_res.text, http_res)
         elif http_res.status_code in [401, 402, 403, 404, 429, 500, 503]: